% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.environment.R
\name{as.environment.packageIQR}
\alias{as.environment.packageIQR}
\title{S3 Method Dispatches for \code{'packageIQR'} Object}
\usage{
\method{as.environment}{packageIQR}(x)
}
\arguments{
\item{x}{\code{'packageIQR'} object, returned value of function \link[utils]{data},}
}
\value{
Function \code{\link[=as.environment.packageIQR]{as.environment.packageIQR()}} returns an \link[base]{environment}.
}
\description{
S3 Method Dispatches for \code{'packageIQR'} Object
}
\details{
Function \code{\link[=as.environment.packageIQR]{as.environment.packageIQR()}} re-runs function \link[utils]{data}
in a \link[base]{new.env}.
}
\examples{
\dontrun{
pkg = setdiff(rownames(installed.packages()), c('rjags', 'VennDiagram'))
# still trying to understand what's wrong with \CRANpkg{VennDiagram}
ev = data(package = pkg) |> as.environment() # not that slow
length(ev)
}
}
\keyword{internal}
