% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_flextable.R
\name{as_flextable.packageIQR}
\alias{as_flextable.packageIQR}
\title{as_flextable.packageIQR}
\usage{
\method{as_flextable}{packageIQR}(x, ...)
}
\arguments{
\item{x}{a \code{'packageIQR'}}

\item{...}{..}
}
\description{
as_flextable.packageIQR
}
