% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrom.R
\name{dataFrom}
\alias{dataFrom}
\title{Data From Package}
\usage{
dataFrom(x = data(package = "datasets"), what)
}
\arguments{
\item{x}{a \code{packageIQR} object, returned value from \link[utils]{data}}

\item{what}{\link[base]{character} scalar or \link[base]{vector}, class(es), see \link[base]{inherits}}
}
\value{
Function \code{\link[=dataFrom]{dataFrom()}} returns a \link[base]{list} of R objects.
}
\description{
Obtain all objects available via \link[utils]{data} from a package
}
\examples{
dataFrom(what = 'data.frame') |> lapply(FUN = head)
dataFrom(what = 'numeric') |> lapply(FUN = head)
data(package = c('datasets', 'MASS')) |> dataFrom(what = 'data.frame') |> lapply(FUN = head)

# vcdExtra::datasets(package = 'datasets') # similar, but not as good!
}
\references{
\url{https://stackoverflow.com/questions/27709936/get-a-list-of-the-data-sets-in-a-particular-package}
}
\keyword{internal}
