% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.R
\name{head.packageIQR}
\alias{head.packageIQR}
\title{\link{head.packageIQR}}
\usage{
\method{head}{packageIQR}(x, n, ...)
}
\arguments{
\item{x}{a \code{'packageIQR'} object}

\item{n}{\link[base]{integer} scalar}

\item{...}{..}
}
\description{
\link{head.packageIQR}
}
