% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split.packageIQR}
\alias{split.packageIQR}
\title{Split \code{'packageIQR'} Object by Package}
\usage{
\method{split}{packageIQR}(x, ...)
}
\arguments{
\item{x}{\code{'packageIQR'} object, returned value of function \link[utils]{data},}

\item{...}{additional parameters, currently not in use}
}
\description{
Split \code{'packageIQR'} Object by Package
}
\details{
Function \code{\link[=split.packageIQR]{split.packageIQR()}} is inspired by function \code{?utils:::print.packageIQR}.
}
\keyword{internal}
